refactor: remove extra folder
