+++ src/components/Footer/Footer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import './Footer.scss';
+import 'components/Footer/Footer.scss';
 
 export default class Footer extends React.Component {
   render() {
