+++ src/components/Header/Header.tsx
@@ -1,7 +1,7 @@
 import NavList from 'components/NavList/NavList';
 import React from 'react';
 
-import './Header.scss';
+import 'components/Header/Header.scss';
 
 export default class Header extends React.Component {
   render() {
