+++ src/components/SearchBar/SearchBar.spec.tsx
