+++ src/index.scss
@@ -1,4 +1,4 @@
-@use './styles' as *;
+@use 'styles/index' as *;
 
 * {
   margin: 0;
